Update fix_fft.cpp
Changed the include for WProgram.h to dynamically select between the old and new (Arduino.h) include based on what version of the Arduino IDE people use.
Also remove the prog_int8_t reference on line 54 and made it int8_t.
Changed the data type of m in fix_fft to be double instead of int in a move to make the dimensions of the LED ranges less restrictive.
Attempted to round and type cast the result of 2^(double m) back to int. But so far it's not playing ball.